+++ src/main/java/com/parse/gwt/client/Parse.java
@@ -236,7 +236,28 @@ public class Parse {
                     });
         }
 
-        public static void requestPasswordReset() {}
+        public static void requestPasswordReset(final AsyncCallback<ParseResponse> callback) {
+            Shape.post(Parse.SERVER_URL + "requestPasswordReset")
+                    .header("X-Parse-Application-Id", X_Parse_Application_Id)
+                    .header("X-Parse-REST-API-Key", X_Parse_REST_API_Key)
+                    .header("X-Parse-Master-Key", X_Parse_Master_Key)
+                    .header("X-Parse-Session-Token", sessionToken)
+                    .asJson(new AsyncCallback<String>() {
+                        @Override
+                        public void onFailure(Throwable throwable) {
+                            callback.onFailure(throwable);
+                        }
+                        @Override
+                        public void onSuccess(String s) {
+                            try {
+                                ParseResponse response = ParseResponse.parse(s);
+                                callback.onSuccess(response);
+                            } catch (Exception e) {
+                                callback.onFailure(e);
+                            }
+                        }
+                    });
+        }
         public static void retrieveUser(String objectId ){}
         public static ParseObject retrieveCurrentUser() {
             Storage storage = Storage.getLocalStorageIfSupported();